Create main.html
Create a html page (named: index.html) with a css style (named: style.css).

It is required to use flexbox. If you use float, this exercise will be considered as failed.

Flexbox is very powerful in order to align responsively elements.

Your web page will be responsive, which means your page will handle a mobile view and a desktop view.